ci: fix branch name to avoid breaking change
It was shortsighted of me that in the revamping of this UI I changed the name of the release given by the workflow. Since this CI deletes the older releases, it means that the current antora-playbook.yaml in the devops-stack will point to ui bundle that no longer exists. This commit will revert that.